+++ calculus.hpp
@@ -27,9 +27,13 @@ public:
     Calculus(std::variant<Scalar, Matrix> const &node_variant, std::map<std::string, std::variant<Scalar, Matrix>> const &node_map);
 
     Scalar Partial(Scalar const &with_respect_to_ptr);
-    Scalar Partial(std::variant<Scalar, Matrix> const &node_variant, Scalar const &with_respect_to_ptr);
+    Matrix Gradient(Scalar const &with_respect_to_11_ptr, Scalar const &with_respect_to_21_ptr, Scalar const &with_respect_to_31_ptr);
+    Scalar Divergence(Scalar const &with_respect_to_11_ptr, Scalar const &with_respect_to_21_ptr, Scalar const &with_respect_to_31_ptr);
+    Matrix Curl(Scalar const &with_respect_to_11_ptr, Scalar const &with_respect_to_21_ptr, Scalar const &with_respect_to_31_ptr);
 
-    Matrix Gradient(std::variant<Scalar, Matrix> const &node_variant, Scalar const &with_respect_to_x_ptr, Scalar const &with_respect_to_y_ptr, Scalar const &with_respect_to_z_ptr);
-    Scalar Divergence(std::variant<Scalar, Matrix> const &node_variant, Scalar const &with_respect_to_x_ptr, Scalar const &with_respect_to_y_ptr, Scalar const &with_respect_to_z_ptr);
-    Matrix Curl(std::variant<Scalar, Matrix> const &node_variant, Scalar const &with_respect_to_x_ptr, Scalar const &with_respect_to_y_ptr, Scalar const &with_respect_to_z_ptr);
+private:
+    Scalar Partial(std::variant<Scalar, Matrix> const &node_variant, Scalar const &with_respect_to_ptr);
+    Matrix Gradient(std::variant<Scalar, Matrix> const &node_variant, Scalar const &with_respect_to_11_ptr, Scalar const &with_respect_to_21_ptr, Scalar const &with_respect_to_31_ptr);
+    Scalar Divergence(std::variant<Scalar, Matrix> const &node_variant, Scalar const &with_respect_to_11_ptr, Scalar const &with_respect_to_21_ptr, Scalar const &with_respect_to_31_ptr);
+    Matrix Curl(std::variant<Scalar, Matrix> const &node_variant, Scalar const &with_respect_to_11_ptr, Scalar const &with_respect_to_21_ptr, Scalar const &with_respect_to_31_ptr);
 };
